+++ backend/fittoring/src/main/java/fittoring/mentoring/presentation/dto/ReviewCreateRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 
 public record ReviewCreateRequest(
-    @Min(1) @Max(5) byte rating,
+    @Min(1) @Max(5) int rating,
     String content
 ) {
